libusiw: Remove unused reliable datagram support
Internally, there was plumbing for reliable datagram support, including
an endpoint table stored as a DPDK rte_hash data structure.  However,
the only transport service exported through the verbs interface is the
standard reliable connected service.

In the future, we intend to move towards a multi-process model.  This
means that every rte_hash data structure must be shared between all
processes.  To reduce the memory overhead in the primary process, we
want to limit these shared data structures as much as possible.

For these reasons, we remove the untested reliable datagram support.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>