libusiw: Eliminate work queue hash tables
In the future, we intend to move towards a multi-process model.  This
means that every rte_hash data structure must be shared between all
processes.  To reduce the memory overhead in the primary process, we
want to limit these shared data structures as much as possible.
Furthermore, storing process-local pointers in a shared memory hash
table will cause problems if other processes see these pointers.
Therefore, we remove the hash table and instead just do a linear scan
for the first matching item.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>