libusiw: Use uthash for per-process queue pair table
In the future, we intend to move towards a multi-process model.  This
means that every rte_hash data structure must be shared between all
processes.  To reduce the memory overhead in the primary process, we
want to limit these shared data structures as much as possible.
Unfortunately, DPDK provides no mechanism to create a process-private
hash table.

We use the BSD licensed uthash library [1] as a replacement, which
provides a lightweight C hash table implementation entirely done with C
preprocessor macros.  This adds a build-time dependency but no
additional dependency at runtime.  The uthash headers are available in
the official package repositories for Fedora, openSUSE, and Ubuntu and
backports are available in EPEL for Red Hat Enterprise Linux and its
rebuilds.

[1] http://troydhanson.github.io/uthash/

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>