Replace build system with autotools
The DPDK build system has many shortcomings that makes it painful to
work with:

 - Explicitly has no support for non-recursive make.

 - Requires a separate Makefile for every target.

 - Very rigid definitions of what an "app" and a "library" are.  Does
   not support linking a library against DPDK, so the resulting apps are
   responsible for linking to every dependent library.  We had worked
   around this by copying a bunch of code from DPDK's build system and
   tweaking it to work for our purposes.

 - Subdirectory support barely works and seems to be almost completely
   untested for external builds.  Lots of subtle bugs that we had to
   work around.

 - Very limited support for "make install".

 - No support for creating a distribution tarball ("make dist" and
   "make distcheck" from autotools).

 - People are more familiar with autotools than the DPDK build system.

While autotools has its own set of disadvantages (in particular, it
generates lots of cruft in the build tree and being somewhat tuned for
GNU extensions), the shortcomings of the DPDK build system outweigh
these disadvantages enough to warrant switching to autotools.  cmake was
also considered but I have more familiarity with autotools so this is
what I am going with.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>