libusiw: Remove usiw_rereg_mr stub, set to NULL
libibverbs 1.2.1 changes the prototype of the ibv_context_ops::rereg_mr
field and also defines the wrapper function ibv_rereg_mr().  The wrapper
function checks if the pointer is NULL and sets errno to ENOSYS in that
case.  Since the wrapper function did not exist previously, just remove
the rereg_mr stub and set the function pointer to NULL, to sidestep this
issue.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>