Create urdmad
Create a urdmad daemon process which sets up the network interfaces and
manages flow director rules.  It communicates with the kernel module via
a character device.  This allows urdma to be used with multiple
processes simultaneously.

urdmad communicates with its subordinate processes using a UNIX domain
socket in order to allocate queue pair numbers and lcores.  This commit
adds a field to the configuration file so that urdmad and its
subordinate processes can find and use the socket.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>