src/libusiw: Pre-allocate all queue pairs
When moving to a multi-process model, we need to be able to allocate
queue pairs to different processes.  The most "fair" way of doing this
is to pre-allocate them and let each process take the queue pairs that
it needs.  Currently, the queue pairs are placed onto a ring and the
verbs code pulls them from the ring as needed.  When this is converted
to a multi-process model, the master process will be responsible for
handing these off, probably via a socket interface.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>