src/liburdma: Do not complete RDMA READ on ACK
If an RDMA READ request finished before the ACK was flushed out, it
would cause a crash:

> ib_read_bw: src/libusiw/interface.c:1510: do_process_ack: Assertion `wqe->state == SEND_WQE_WAIT' failed.

This crash occurred because we were trying to complete the RDMA READ
request simply because it was ACK'd.  We need to treat RDMA READ
requests specially, and *not* complete them when they are ACK'd, because
we haven't received all of the data yet.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>