liburdma: Set MAC address and bring interface up in one step
This ensures that the IB device is registered with the correct MAC
address (previously, it was registered with a "fake" GUID and then the
GUID was updated once the MAC was changed later).

The original behavior was due to a misfeature in the KNI kernel module
in versions of DPDK prior to 16.07, which reset the MAC address when the
interface was brought up.  Bump the minimum requirement to DPDK 16.07
accordingly.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>