kmod: Clean up node_guid, sys_image_guid, port_gid
These are now all the same.  sys_image_guid is hardcoded to be the same
as node_guid.  port_gid is returned from the netdev, but should always
match node_guid due to the net handler.  Always adjust the node_guid on
every MAC address change, even before the device is registered, to
ensure consistency.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>