configure.ac: Set MACHINE_CFLAGS using DPDK build system output
We had defaulted to using -march=native which does ensure that gcc
doesn't error out on header files; however, we really should match the
same MACHINE_CFLAGS that were used to compile DPDK, to avoid having any
issues due to differences in the machine-dependent code.  We use make to
invoke the DPDK build system to extract the MACHINE_CFLAGS.

Instead of allowing the user to override these flags via the
MACHINE_CFLAGS environment variable, we cache the value, which can be
overriden by editing the config.cache file.  Most users should not need
to override this variable; if they do, it should be considered a bug.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>