Fix IFF_UP configure check
Only define the AVOID_NET_IF_H symbol if the condition is actually true,
as opposed to defining it with the value 0.  Cache this check using
AC_CACHE_CHECK, which actually reduces the amount of code in
configure.ac.  Add a check that IFF_UP is actually defined by net/if.h.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>