Fix linking on Ubuntu
A previous commit resulted in DPDK_LIBS being unconditionally defined to
the empty string.  Ubuntu's linker requires that libraries occur after
their dependencies on the linker command line, so that broke Ubuntu.
Move the -ldpdk into DPDK_LIBS so that it always gets included at the
end of the linker command line.

Fixes: da7967fdfded ("liburdma: Set MAC address and bring interface up in one step")
Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>