liburdma: Remove pointer from device to context
Instead of looping over ibv_devices in the main progress thread loop,
loop over ibv_contexts, thus removing the pointer from the device to the
context.  This makes more sense because there can be multiple contexts
open for a single device.

This also fixes a long-standing bug where we leaked a dangling pointer
to a context after the context was closed with ibv_close_device(),
leading to a segmentation fault in the progress thread.  This issue was
hidden for a long time since the RDMA CM creates long-lived contexts
that applications should use and never closes them, but openmpi creates
and closes multiple contexts during its execution.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>