liburdma: Clear wc_flags field of work completions
openmpi uses the wc_flags to determine whether or not there is immediate
data attached to messages.  We did not fill in the wc_flags field, so
openmpi would read the garbage in that field, conclude that there was
attached immediate data, and try to interpret the garbage in that field.

By zeroing out the wc_flags field, we prevent this from happening.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>