liburdma: Manipulate driver context list in progress thread only
Commit 9de37f6029bc ("liburdma: Remove pointer from device to context")
introduced a race condition when opening a new context with
ibv_open_device(), where the user thread would add an entry to the
driver context list but the progress thread would never see the added
context.  This is because the sys/queue.h interface is not thread-safe.
This worked prior to this commit because the set of drivers was static
and finalized before the progress thread starts, whereas contexts are
opened/destroyed by the user application outside of our control.

To fix this, we must use alternate mechanisms to inform the progress
thread of new and removed contexts.  We use a DPDK ring to inform the
progress thread that a context has been added, and new contexts are
added prior to the next iteration of the main loop.

Handling closed (destroyed) contexts is challenging, because
ibv_close_device() frees the context structure, but we can't directly
remove the context from the progress thread's list since the list is not
thread safe.  To solve this problem, we make the progress thread's
context list indirect; the list consists of handles which contain an
atomic pointer to the context.  When the context is destroyed with
ibv_close_device(), we atomically set the pointer to NULL, and then the
progress thread removes the handle from the list and destroys it the next
time it goes around the loop.

Fixes: 9de37f6029bc ("liburdma: Remove pointer from device to context")
Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>