liburdma, urdmad: Use consistent value for descriptor limits
When splitting urdmad from liburdma, the logic to get the descriptor
limits was not moved to urdmad, leading to a discrepency between the
descriptor count that liburdma was assuming and the actual ring and
mempool sizes.

Set this in urdmad when setting up the queues and then transfer this
information to liburdma for each initialized queue pair.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>