liburdma: Fix thread safety of CQ rings
The CQ can be polled by multiple threads at the same time, so the
appropriate thread safety measures must be taken.  Previously, we had a
lock around the CQ operations, such that only a single thread could
access the CQ rings at a time.  When we removed the lock, we did not
change the flags appropriately.  This caused a potential race in
applications which polled the CQ from multiple threads.

This is fixed by removing the RING_F_SC_DEQ flag from the cqe_ring of
ready CQEs, and removing the RING_F_SP_ENQ flag from the free_ring.
This ensures that multiple threads can poll the CQ and remove events
from it, while keeping the single-thread optimization for the progress
thread which is the only thread that can remove entries from the
free_ring and add them to the cqe_ring.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>