urdmad: Immediately exit on port_init() failure
The urdmad service is allowed to output to stderr/syslog, since it is an
application and not a library. There is no recovery action that we can
take if any of the setup steps fail in usiw_port_init(). Thus, don't
return an error code but instead exit the application after printing an
appropriate message to stderr.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>