liburdma,kmod: Fix ird and ord negotiation
The RDMA CM inexplicably and silently reverses the meaning of ord and
ird (i.e., initiator_depth and responder_resources) on the server side
when it processes a connection request event from our driver.  This
means that from the user's perspective, the meaning of these variables
is always from the client's perspective.  Since this is confusing, keep
the meaning within the driver to match the local side, but do the
appropriate swap to undo the swap done by the RDMA CM in the server.

This additionally fixes a few other places in the code where this was
reversed incorrectly or the wrong set of fields was looked up as a
result of the above, so now ird/ord negotiations should work as
expected.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>