liburdma: Ensure correct message ordering
Previously, we would only dequeue recv WQEs as incoming packets arrived.
However, this could cause out-of-order message delivery if a packet was
placed out of order.  Now, we move WQEs from the ring to our progress
thread linked list immediately before we process a burst of received
packets, so that the ordering is properly maintained.

Additionally, we no longer post the completion if missing frames.  While
we can place the data, we must wait for any prior RDMA WRITE requests or
RDMA READ response packets before delivering the completion to the
application.  When we complete a message and have no missing
intermediate frames, we walk the receive queue starting from the head to
make sure that we report all pending completions in the correct order.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>