kmod: Fix lock release on chardev read() with bad address
Sparse complains if a lock is released by a function without having been
previously acquired by that same function.  Annotate that we release the
lock without acquiring it (i.e., the caller acquires it before calling
us).

This reveals that in the case where the user passes an invalid user
address to read(), we fail to unlock the lock or release references that
we have acquired.  Fix this.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>