kmod: Implement disassociate_ucontext
When a network device is removed from the system, the netdev device
notifier callbacks are invoked to inform the system that the device will
be unregistered. In response, we unregister the corresponding verbs
device.

If urdmad is killed while a verbs application is using the corresponding
verbs device, ib_uverbs_remove_one will block until the verbs
application exits. This in turn blocks other processes accessing the
network stack.

The new disassociate_ucontext handler currently does nothing, but when
present, the IB core does not block in ib_uverbs_remove_one, and thus no
longer blocks the netdev device notifier callback, preventing a hang.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>