kmod: Do not disable interrupts when locking chardev
On urdma_chardev_release(), we potentially free a CEP if we were the
last remaining reference to it. This in turn locks the CEP using
spin_lock_bh/spin_unlock_bh, which must not be called with interrupts
disabled.  Switch to spin_lock_bh() for the file locks since we don't
really need to disable interrupts.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>