Run urdmad as normal user
Put the socket file in $XDG_RUNTIME_DIR by default so we can run this as
an arbitrary user using the systemd user service manager (systemd
automatically sets the $XDG_RUNTIME_DIR environment variable before
running urdmad or any user login session). This is required for openmpi
>= 2.0, which does not support running "sudo" underneath it.  To support
this, the "socket" attribute in the configuration file has been made
optional.

On deployed systems, this change requires removing the urdmad service
file from /usr/lib/systemd/system and putting the new service file under
/usr/lib/systemd/user (or equivalent). The service can then by run as any
user in the "dpdk" group after permissions have been set up
appropriately. There is no more need for a dedicated dpdk user; however,
any verbs applications must be run as the same user as urdmad (this is a
DPDK limitation).

This commit also adds the hostname to the DPDK file-prefix; this is
required on environments with NFS-shared home directories in which urdma
may be run on multiple systems at once.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>