util: Add binary heap implementation and tests
This binary heap implementation will be used to track incoming packets
which mark the end of a message which has been placed but cannot yet be
delivered according to the iWARP specification. When the corresponding
messages may be delivered, we use the binary heap to deliver them in
message order even if the packets were originally received out of order.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>