urdmad: Allow specifying devices by PCI address
Make configurations less ambiguous by allowing the user to request a
device with a specific PCI address instead of relying on the port index.
If the user does not supply a PCI address, the configuration behaves the
same way as before, although users are recommended to add PCI addresses
to their configuration to ensure that they are using the devices that
they think they are.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>