urdmad: Implement ntuple filter support
Currently the only filter type that we support is Flow Director which is
only supported on Intel NICs, since it is the only perfect-match filter
type supported by the i40e driver. This commit adds support for the more
generic ntuple filter type as well, which is supported by at least the
ixgbe and e1000 drivers.

Prefer ntuple to flow director when it is available since ntuple is far
simpler. Prefer 5tuple to 2tuple, but use whichever the NIC supports if
it only supports one or the other.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>