src/kmod: Fix dma_device handling in Linux 4.11+
In Linux 4.11, the struct ib_dma_mapping_ops was folded into the
core DMA ops, so we no longer need to maintain a separate struct for
that. In addition, the dma_device field is no longer managed by the
hardware driver; instead, we should set dev.parent and the RDMA core
will manage the dma_device field for us automatically.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>