urdmad,liburdma: Change QP state lock from spinlock to mutex
Spinlocks are not safe to use when a process may block or execute a
system call because the waiting process may consume CPU cycles
indefinitely. Change this to a mutex, which on modern systems behaves as
a spinlock in the absence of contention, but blocks when there is
contention.

Add -pthread to CFLAGS so that everything is compiled with thread
support on Linux.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>