liburdma: Hold conn_event_lock while destroying QP
There is a race between the progress thread trying to transiction the
queue pair to the error state and a user trying to destroy the queue
pair.  Destroying the queue pair in the kernel requires an exclusive
lock on the queue pair uobject which is tried only once; if another
thread happens to be calling another uverbs function at the same time,
the exclusive lock fails and the operation fails with EBUSY.

Eliminate this race condition by holding a lock such that we prevent
these two functions from being called simultaneously.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>