Convert from sys/queue.h to CCAN list/list.h
The rdma-core project uses the list implementation from CCAN instead of
the non-portable sys/queue.h. Both implemenations define macros with the
same name, making them incompatible with each other. Since we will need
to work with rdma-core, import and use the CCAN list implementation. The
recommendation for CCAN code is to include the code from ccan directly
into a project's source directory under a ccan/ directory, so do that.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>