Add list_test program to build system
There has been a list test program from the beginning of the project;
move it to our new tests directory and add it to the Makefile.am so it
gets run with a "make check" invocation. This will be used to verify
that switching from sys/queue.h to the CCAN list implementation does not
break list handling.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>