Implement RFC 7306 atomic mask fields
For completeness, implement the mask fields that RFC 7306 requires
responders to support. Note that no Linux RDMA verbs application can
ever send these requests with masked fields because the rdma-core API
does not provide a way for applications to set the masks. Thus, we use
an optimized code path in the common case where the masks are 0.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>