Add version field to the urdmad socket protocol
To ensure compatibility, add a version field to the socket protocol for
urdmad. This will be used to ensure that a urdma process cannot connect
to an incompatible urdmad instance.

Old urdma client and server versions will fill the space now used by the
version field with 0, as the maximum number of lcores that DPDK can use
is limited to 127 by default. This ensures that we do not attempt to
parse an older version of the protocol format.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>